requirements.txt
torch
torchvision
torchaudio
wandb
pillow
numpy
tqdm